Added Bluetooth Native Modules
Added a BLE module and package that will enable one to scan and connect to a peripheral device. It also allows a peripheral to broadcast. Another file that was added was a util file blePermissions.ts that will be used to manage all the permissions required for BLE to work seamlessly. What is left to add is a file that requests for those permissions.